move `FakeView` classes into `Statamic\Testing`
They're concrete classes, not traits, so they don't belong in `Concerns`.
